backend/requirements.txt
fastapi
uvicorn[standard]
google-genai
google-adk
python-dotenv
python-multipart
pydantic
pillow
aiofiles
